Add support for Postgresql as well as sqlite

+++ db/migrations/schema.sql
@@ -1,13 +1,13 @@
 create table if not exists feeds
 (
     feed_name     text primary key,
-    latest_cursor integer
+    latest_cursor bigint
 );
 
 create table if not exists feed_posts
 (
     feed_name text    not null,
-    time_us   integer not null,
+    time_us   bigint  not null,
     did       text    not null,
     rkey      text    not null
 );
